Parsing & restructuring
The parser is slowly taking shape, but it's still unclear exactly what's
the most ergonomic way to structure it. Many failed implementations have
been attempted which proved to be difficult to work with or difficult to
understand. Attempting to move towards TryFrom implementations for most
types and having parsing work by calling try_from on each incremental
chunk to determine what is valid and what isn't. Not yet functional.