Tests: macros to simplify testing
Implemented some macros to reduce the massive amount of boilerplate
required to perform tests. Also added try_from_iter which is a function
which will try to create an EffectiveAddress from an iterator over
operators and operands. This is not ideal since it's not an implementation
of TryFrom, which was attempted but failed due to an overlapping
implementation from core. Will need to revisit later. Currently not
used, but likely will help further simplify testing, allowing even
shorter constructions of EffectiveAddresses.