CPU: implemented some ADD instructions
Unsure if the manner in which unsigned registers are interacting with
signed immediate values is correct. As far as I understand, registers
are unsigned, but instructions may treat them as signed for the purpose
of operating on them. Perhaps I should be casting the register to a
signed int rather than casting the immediate to an unsigned int. Unsure,
I will find out when I write tests.