CPU/Registers: restructure to allow instructions to work.
Prior to this commit there was no easy method through which register
values could be updated if you had a variant of the Register enum and
wanted to act upon that register. Methods have been added to get and set
register values based on variants for each of the various register
sizes. It is not clear if this approach will remain in place as it is
extremely verbose and repeats the various register variants twice. This
is not ideal but works, so it will work for now. I strongly suspect
there is a much more efficient way to achieve this, possibly through
traits.