Parity flag usage and add SIB module into tree.
The parity flag is set if the least significant bit (LSB) of the result
of the last arithmetic operation contained an even number of high bits.
Have added a method on eflags to compute this and update the flag
accordingly. Have added it to the various ADD instructions that have
been implemented thus far so that they now update the parity bit.