Traits: making adding generic
Using the num crate's wrapping_add and a custom least_significant_byte
trait, we can now have a generic add function on the CPU. This is
helpful to reduce the amount of repetitive functions and function calls
to set the parity bit. The parity bit is now set in the CPU's add method
and any instruction that needs to perform addition can call that method
and automatically have flags set.