Alternate CPU function architecture.
Continually unwrapping operands that have already been validated to be
correct previously within each CPU function is not great. This code does
not work, and will likely not be followed through with, but is being
committed for future reference.

The main issues with the new approach contained within is that is
doesn't fix the issue as much as it moves it. You still need to perform
unwrapping on an instruction-by-instruction basis somewhere.