Register enum is now composed rather than repeated
Prior to this there were the following separate enums: Register,
Register32, Register16, and Register8. This was unnecessarily
repetitive, so the Register enum now has variants for each size and then
simply wraps them. The From implementations have also been adjusted to
make this as ergonomic as possible, although some repetition is
unfortunately unavoidable as enum variants currently cannot be
considered as independent types.