Add Emulator, Ram, and RegisterOrMemory
The Emulator struct has been added as a container for all components.
A basic Ram struct has been added which currently is just a wrapper for
a Vec<u8>, however this will likely also contain further information in
the future. RegisterOrMemory is an enum which, as the name implies,
contains either a register or memory (effective address). This is useful
for instructions which can either take a register or memory reference.
This paves the way for hopefully better instruction operand parsing,
whereby we can unwrap into specific types based on what is required in a
particular instruction (Cpu method).