RegisterOrMemory: conversion from Operand and associated Error
Added a TryFrom<Operand> for RegisterOrMemory. If the Operand is an
Immediate, then an error must be thrown, for which a new error has been
added. This leverages the previous commit, whereby you can now take
ownership of the fields contained within an Operand, which is required
here to convert the Operand into a RegisterOrMemory without cloning.