Cpu and eflags: flag computation and bitwise and
Implemented bitwise and in a generic manner using the num_traits crate.
Realised that we do not need the num crate, only the num_traits crate
which is was re-exporting. We also don't need the LeastSignificantByte
trait by leveraging the correct traits. Have implemented computation for
zero and sign flags and used these to implement a generic bitwise and
method on Cpu, which has been used to implement 3 new and instructions.