Registers: carry flag fix casting to unsigned for addition
If a 64-bit unsigned value was passed in on a 64-bit machine, this would
be converted into an isize, which would be equivalent to an i64 and
fail. This would be even worse on a 32-bit machine. We now ensure to
cast to a value that is double the size of the one passed in, to ensure
that the largest unsigned value possible is still covered within the
signed range. This is very hacky and should be changed. All that really
needs to be done is reinterpreting the bits as an unsigned integer of
the same size as the one passed in.