Register: parity flag setting
I have realised that two's complement addition/subtraction works
identically whether it's signed or unsigned. That is to say that it's
only a matter of how you interpret the result, the actual operations are
identical. Knowing this, we can always treat everything as unsigned, and
then interpret it as signed or unsigned depending on what we need. This
is obvious in hindsight and many things pointed to this, but it
definitely simplifies the design.

There is still some residual awkwardness to sort out. For instance, some
functions have trait bounds of `Unsigned + AsUnsigned`. This is
nonsensical, but is required because the calling functions require both.
In order to fix this we will need to universally require the same trait
bound. It would be nicer if we could take `AsUnsigned` everywhere, so
that all the callers don't have to repeatedly call `.as_unsigned()`.