Register: compute auxiliary carry flag correctly
We calculate the auxiliary carry flag differently based on whether it's
addition or subtraction. During addition, if the addition of the lowest
nibbles results in the bit above the first nibble (5th bit) being 1, we
set AF. During subtraction, if the subtraction of the lowest nibbles
results in an underflow, this indicates that a borrow would have been
generated. Therefore, we check if the operand being subtracted is
greater than the operand being subtracted from. If so, the auxiliary
carry flag is set. I believe this should work, however am not entirely
positive and will have to run some tests on a real processor to verify.