CPU sub: tests
No changes needed to the flag setting computation, everything worked
first try luckily. Not all permutations of flags are covered, but
according to the truth table on Stack Overflow, these appear to be the
only permutations which can actually result in practice. May be of
interest looking further into this, but all of these work.