Eflags tests + AsSigned trait
Implemented tests for a few additional eflags compute methods. There is
some inconsistency between what arguments are taken for each eflags
compute method, which can likely be evened out by having an AsSigned
trait, as we have an AsUnsigned trait. This trait has therefore been
implemented, but is not in use yet. The down side to using it, is that
some eflags compute methods will likely need to re-compute the
calculation that led to the result, rather than having the result passed
in. This is inefficient and it's unclear whether it's worth going down
this path. Regardless, AsSigned will almost certainly be needed later
when we want to interpret values as such.