Eflags: infer carry flag rather than repeating operation
We infer whether a carry was generated rather than repeating the
addition/subtraction and checking that we do not wrap. This is more in
line with how we infer the overflow flag. This is simpler because we
don't need to special case the carry flag, we simply look at the input
and output and infer whether it wrapped. This is less generic than
recomputing the operation and checking for wrapping, however as long as
the carry cannot be more than 1 the logic should be correct. This should
always be the case in x86.