Instruction: restructure to better match real architecture.
Currently we do not parse instructions in the same manner as they are in
real x86 (with SIB, ModRM, opcodes, etc.). This works and is simpler to
implement, however misses the goal of mirroring exactly how it works.
This will require sweeping change throughout almost every section of the
code base, so it is being implemented separately first so that it may
live dormant until ready to be switched out.