Restructure: remove unnecessary files and rename others
Old SIB file was no longer needed has been replaced by the
newinstruction version. The new instruction format which matches the
actual assembled instruction format more closely has been placed in the
new encodedinstruction module, since that seems to appropriately capture
its intent for now. It is not yet needed, so can be updated later when
the context for its existence is better known. The thinking at this
moment in time is that this version of the instruction will be used for
assembly/disassembly. So while it may not be the most ergonomic format
to work with when actually extracting details such as operands, it's
closest to the instruction format of actual machine code. How I imagine
that this will work is that when assembling or disassembling,
conversions will need to take place between this machine code version
and the more ergonomic version that holds the semantic details of the
instruction.