Instruction: parsing, representation and tests
In the process of implementing memory address resolution (which is still
not fully complete), shortcomings were discovered and mostly addressed. Firstly,
immediate values are now u32 as they should be. This required changes to
many sections of code that relied on it not being so. Created and implemented
a trait for reading and writing to registers. This allows values to be
read in a more similar manner to immediate rm values. Unclear if it's
worth it, but it seems to make some things more consistent. Tests have
been added to cover size inference and other tests have been adjusted to
account for these changes. All are passing except for one, which is
entirely due to new test cases being added that will be addressed soon.
The manner in which we handle signed immediate values has not been fully
threshed out and these tests serve as a reminder of those shortcomings.