Instruction: validation and tests
Tests that were previously failing have now been fixed. The issue was
that all values, including immediates, are taken as being unsigned. If a
value is to be singed, that is simply a matter of interpretation, but we
do not encode that anywhere. Therefore, -1 is the same as the maximum
unsigned value. This caused problems because the validation would ensure
that immediates passed to instructions were of the expected size. This
is true if a size directive is provided, such that if you pass the wrong
size directive, it is rejected, however it's not true otherwise.

Prior to this change, if a size directive was not provided, it would be
inferred, and validated. Therefore, in the case that a BYTE was expected
and you passed -1, this would be 0xFFFFFFFF (max u32) and consequently
be inferred to be a DWORD. It would therefore be rejected, even though
0xFF could equivalently be used to express -1 in BYTE form.

The change is to therefore not attempt to infer sizes and accept
anything as long as a size directive is not provided. The value will
simply be truncated later on if it's too large. This has been validated
using NASM.