Cpu: cleanup and unwraps rather than ignoring Results
Now that reading/writing to memory has been implemented, the calls have
become fallible and hence return results. At the point of calling these
functions when executing instructions in the CPU, there is nothing that
can be done and the instruction should fail if the write fails. As such,
the result values are now all unwrapped to force a panic in the event
that a memory write has failed.

Also cleaned up some unused imports.