Cpu/Instruction: MOV and change cpu function signature
CPU functions now take an Operands struct (just a Vec<Operand>), which
is clearer than the previous signature which took a &Instruction,
because that also contained the mnemonic and a link to the CPU function
(itself).

MOV has also been implemented, along with tests. Only the most common
MOV instructions have been implemented so far. There are still other
less common variants which are pending.