CPU: Remove unnecessary operand unwraps
Further to the previous commit, we can actually entirely remove the
unwrap_operands macro call since we do nothing with the values we
unwrap. An argument could be made that we should verify that we have
indeed been passed the correct operand, but it does not seem to be the
responsibility of the callee to check this.